move form and list component to app-configure
